Remove a bunch of no-longer-needed code, and start fixing tests
