+++ webapp/apps/taxbrain/tests/test_models.py
@@ -89,3 +89,34 @@ class TestTaxBrainStaticFields(TaxBrainFieldsTest):
                                   use_puf_not_cps=False)
 
         assert not model.use_puf_not_cps
+
+    def test_set_fields(self, test_coverage_behavioral_gui_fields):
+        start_year = 2017
+        fields = test_coverage_behavioral_gui_fields.copy()
+        fields['first_year'] = start_year
+        self.parse_fields(start_year, fields)
+
+    def test_data_source_puf(self, test_coverage_behavioral_gui_fields):
+        start_year = 2017
+        fields = test_coverage_behavioral_gui_fields.copy()
+        fields['first_year'] = start_year
+        fields['data_source'] = 'PUF'
+        model = self.parse_fields(start_year, fields, use_puf_not_cps=True)
+
+        assert model.use_puf_not_cps
+
+    def test_get_model_specs_with_errors(self, test_coverage_behavioral_gui_fields):
+        start_year = 2017
+        fields = test_coverage_behavioral_gui_fields.copy()
+        fields['BE_sub'] = [-0.8]
+        fields['BE_inc'] = [0.2]
+        fields['first_year'] = start_year
+        fields['data_source'] = 'PUF'
+        model = self.parse_fields(start_year, fields, use_puf_not_cps=True)
+        (reform_dict, assumptions_dict, reform_text, assumptions_text,
+            errors_warnings) = model.get_model_specs()
+
+        assert len(errors_warnings['behavior']['errors']) > 0
+        assert len(errors_warnings['behavior']['warnings']) == 0
+        assert len(errors_warnings['policy']['errors']) == 0
+        assert len(errors_warnings['policy']['warnings']) == 0
